+++ src/pmutils.jl
@@ -70,7 +70,6 @@ function tvstm(A::PM, tf::Real, t0::Real = 0; solver = "auto", reltol = 1e-3, ab
       function update_func!(A,u,p,t)
          A .= PeriodicMatrices.tpmeval(p,t)
       end
-      #DEop = DiffEqArrayOperator(ones(T,n,n),update_func=update_func!)  
       DEop = MatrixOperator(ones(T,n,n);update_func!)     
       prob = ODEProblem(DEop, u0, tspan, A)
       sol = solve(prob,MagnusGL6(), dt = dt, save_everystep = false)